requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Image processing
opencv-python>=4.8.0
scikit-image>=0.21.0
tifffile>=2023.7.10
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0

# Cellpose dependencies
torch>=2.0.0
torchvision>=0.15.0

# HTTP requests
requests>=2.31.0

# File handling
natsort>=8.4.0

# Optional: for better performance
# psycopg2-binary>=2.9.0  # for PostgreSQL
# redis>=4.6.0  # for caching/task queue

